🛠️ Bootstrap project scaffolding and fix cross-platform build bug 🛠️
Ran the bootstrap workflow against the existing AI.md/IDEA.md spec: added
missing PART-1-required root files, added the Claude Code rules loader
set, fixed a doc/URL inconsistency, and fixed a freebsd build failure
uncovered while verifying the Docker build.

- CLAUDE.md: add as a short loader pointing to AI.md/IDEA.md (required
  by AI.md PART 1 § Allowed Root Files, was missing)
- .claude/rules/: add the 13 topic rule files listed in AI.md PART 0
  (ai-rules, project-rules, config-rules, binary-rules, backend-rules,
  api-rules, frontend-rules, features-rules, service-rules,
  makefile-rules, docker-rules, cicd-rules, testing-rules), each a
  short loader referencing the relevant AI.md PART numbers
- README.md: replace `http://localhost:PORT` examples with the
  documented `official_site` (`https://scour.li`) per AI.md PART 1 §
  URL Standards; correct the operator-access note — config reloads via
  a file watcher, not SIGHUP (src/signal ignores SIGHUP intentionally)
- src/backup/diskspace_unix.go: fix `go build` failure on
  freebsd/amd64+arm64 (`stat.Bavail * uint64(stat.Bsize)` mismatched
  int64/uint64) by explicitly casting both `stat.Blocks` and
  `stat.Bavail` to uint64; verified full `make build` matrix
  (linux/darwin/windows/freebsd amd64+arm64, both binaries) now exits 0
- TODO.AI.md: add two outstanding items — Gitea/Forgejo workflow parity
  for daily.yml/docker.yml, and a decision on adopting beta.yml (no
  beta branch exists yet)

Verified: `make build` exits 0 for all 8 platform targets and both
binaries; `make test` passes with 80.0% coverage; go-lint reports no
convention violations.

+++ src/backup/diskspace_unix.go
@@ -12,8 +12,8 @@ func DiskUsage(path string) (used, total uint64, err error) {
 		return 0, 0, err
 	}
 
-	total = stat.Blocks * uint64(stat.Bsize)
-	free := stat.Bavail * uint64(stat.Bsize)
+	total = uint64(stat.Blocks) * uint64(stat.Bsize)
+	free := uint64(stat.Bavail) * uint64(stat.Bsize)
 	used = total - free
 
 	return used, total, nil
